Start outsourcing the webxr API
