+++ lib/src/threejs/interop/base.dart
@@ -3,7 +3,7 @@ library base;
 
 import 'dart:html' as html;
 
-// import 'package:flutter_web_xr/src/webxr/core.dart';
+import 'package:flutter_web_xr/src/webxr/interop/core.dart';
 import 'package:flutter_web_xr/utils.dart';
 import 'package:js/js.dart';
 
@@ -17,7 +17,7 @@ class WebGLRenderer {
   external factory WebGLRenderer([Object options]);
   external html.CanvasElement get domElement;
   external void render(dynamic scene, dynamic camera);
-  // external XR get xr;
+  external XR get xr;
   external void setSize(num width, num height);
   external set autoClear(bool value);
   external clear([bool color, bool depth, bool stencil]);
