+++ lib/src/webxr/interop/core.dart
@@ -1,17 +1,5 @@
-import 'package:flutter_web_xr/web_xr_manager.dart';
+import 'package:flutter_web_xr/src/webxr/interop/xr_session.dart';
 import 'package:js/js.dart';
-
-// @JS()
-// class XR {
-//   external bool get enabled;
-//   external set enabled(bool value);
-
-//   external bool get isPresenting;
-//   external set isPresenting(bool value);
-
-//   external XRSession getSession();
-//   external dynamic getController(num index);
-// }
 
 @JS('navigator.xr')
 external XRSystem? get xrSystem;
@@ -26,3 +14,16 @@ class XRSystem {
 
 @JS('navigator.xr.requestSession')
 external dynamic startSession(String sessionType);
+
+// not used currently
+@JS()
+class XR {
+  external bool get enabled;
+  external set enabled(bool value);
+
+  external bool get isPresenting;
+  external set isPresenting(bool value);
+
+  external XRSession getSession();
+  external dynamic getController(num index);
+}
