+++ lib/src/webxr/models/web_xr_manager.dart
@@ -1,0 +1,1 @@
+class WebXrManager {}
