--- lib/xr_session.dart
@@ -1,38 +1,0 @@
-// import 'package:flutter_web_xr/web_xr_manager.dart';
-
-// class XRSession {
-//   String? depthDataFormat;
-//   String? depthUsage;
-//   String? domOverlayState;
-//   String? environmentBlendMode;
-//   dynamic
-//       inputSources; // Für dieses Beispiel habe ich angenommen, dass die inputSources eine Liste von dynamischen Objekten sind
-//   String? interactionMode;
-//   String? preferredReflectionFormat;
-//   // XRRenderState? renderState;
-//   dynamic renderState;
-//   String? visibilityState;
-
-//   XRSession({
-//     this.depthDataFormat,
-//     this.depthUsage,
-//     this.domOverlayState,
-//     this.environmentBlendMode,
-//     this.inputSources,
-//     this.interactionMode,
-//     this.preferredReflectionFormat,
-//     this.renderState,
-//     this.visibilityState,
-//   });
-
-//   @override
-//   String toString() {
-//     return 'XRSession(depthDataFormat: $depthDataFormat, depthUsage: $depthUsage, domOverlayState: $domOverlayState, environmentBlendMode: $environmentBlendMode, inputSources: $inputSources, interactionMode: $interactionMode, preferredReflectionFormat: $preferredReflectionFormat, renderState: $renderState, visibilityState: $visibilityState)';
-//   }
-
-//   // external void updateRenderState(XRRenderStateInit state);
-// }
-
-// class XRRenderState {
-//   // Hier können Sie die Eigenschaften von XRRenderState hinzufügen
-// }
